Groundwork for the second version of inlining
